Fix processing the kernelcache of iOS >13.2
Since iOS 13.2, the kernelcache ASN.1 structure has a sequence of two
integers following the kernel data. However, getRawKeybag was *assuming*
that if there was anything after the data blob then the file was encrypted,
and tried to parse that next node as a keybag, without even checking if it
was an asn1 octet string. This made the script crash on >13.2.

I'm now checking that the next node is actually an octet string before
treating it as a kbag:
- If there's no next node, the file isn't encrypted.
- If there is a next node and it is an octet string, the file is encrypted
  and the octet string is parsed as a keybag.
- If there is a next node but it's another type, the file isn't encrypted.