Add a way to override the version number
When dumping keys of a beta version, the Version will show up as eg. "13.0"
instead of "13.0 beta 5", and as far as I know the beta number isn't
present in any manifest we can look at in the ipsw. Instead, this commit
adds a way to override the version, by passing a text file mapping build
IDs to version numbers.

For example, put this in a versions.txt:

17A5821e 13.1 beta
17A5831c 13.1 beta 2
17A5837a 13.1 beta 3
17A5844a 13.1 beta 4

Then pass --version-map versions.txt to ipsw_keys.py.